:recycle: desacoplate logic for function update
