+++ src/services/user.service.ts
@@ -1,5 +1,6 @@
 import {randomUUID} from "crypto";
 import dynamodb from "../utils/Dynamodb";
+import responseObject from "../utils/Response";
 
 export default class UserService {
     public static create = async (data: any) => {
@@ -9,7 +10,7 @@ export default class UserService {
             Item: { ...data, pk: id },
         };
         await dynamodb.put(params).promise();
-        return params.Item;
+        return responseObject(201, params.Item);
     }
 
     public static findById = async (userId: string) => {
@@ -19,6 +20,23 @@ export default class UserService {
             TableName: "users",
         };
         return await dynamodb.query(params).promise();
+    }
+
+    public static update = async (data: any, userId: string) => {
+        const userFound = await UserService.findById(userId);
+        if (userFound.Count === 1){
+            const params = {
+                TableName: "users",
+                Key: { pk: userId },
+                UpdateExpression: "set #name = :name",
+                ExpressionAttributeNames: { "#name": "name" },
+                ExpressionAttributeValues: { ":name": data.name },
+                ReturnValues: "ALL_NEW",
+            };
+            const result: any = await dynamodb.update(params).promise();
+            return responseObject(200, {message: "User updated!", result});
+        }
+        return responseObject(404, {message: "User not found!", userId});
     }
 
 }
